python-app/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.3 # use pre-compiled version for container
requests==2.32.3